Using the JSON
Implementing the JSON file in that loading characters and enemies is now possible. Henchamn still have a different way of being implemented for ease of access. Future versions will have to be managed a bit better, the two loop system has outlived it's usefulness and now it should be combined into one.