Update 5/3
Changed a bit of the orginazation a little bit more, now have a random project about WorkTimer to try and implement something random I thought of at work, but also continued updated the website to get it ready for hosting. The look of the website is pretty far out but the functionality is almost at where it can be before implementing a database!