Data redundancy! UI Changes! More Functionality!
We've got a basic AWS bucket set up now to allow users to have a JSON file they can write characters to in order to retrieve later. We'll be highly customizing this soon. Only just now figured out a semi secure way to upload this without getting hacked. Slight issues with the main.js that won't let it be uploaded still though